# openapi.yaml
openapi: 3.0.3
info:
  title: API Penjualan
  version: 1.0.0
  license:
    name: Academic use
  description: |
    Kontrak REST API untuk mengelola resource penjualan pada database Neon.
    Nilai `total` dihitung oleh server dari harga produk dikali jumlah barang.
servers:
  - url: http://localhost:3000
    description: Server lokal
  - url: https://your-app.vercel.app
    description: Ganti dengan URL deployment
tags:
  - name: Penjualan
    description: Operasi data dan status transaksi penjualan
security: []
paths:
  /api/penjualan:
    get:
      tags: [Penjualan]
      summary: Ambil daftar penjualan
      operationId: listPenjualan
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/StatusFilter'
      responses:
        '200':
          description: Daftar penjualan berhasil diambil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PenjualanListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags: [Penjualan]
      summary: Tambah penjualan baru
      description: Transaksi baru selalu dibuat dengan status `pending`.
      operationId: createPenjualan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePenjualanRequest'
      responses:
        '201':
          description: Penjualan berhasil dibuat
          headers:
            Location:
              description: Lokasi resource yang baru dibuat
              schema:
                type: string
                example: /api/penjualan/11
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PenjualanResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/penjualan/{id}:
    parameters:
      - $ref: '#/components/parameters/PenjualanId'
    get:
      tags: [Penjualan]
      summary: Ambil satu penjualan berdasarkan ID
      operationId: getPenjualanById
      responses:
        '200':
          description: Detail penjualan berhasil diambil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PenjualanResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      tags: [Penjualan]
      summary: Hapus penjualan
      operationId: deletePenjualan
      responses:
        '204':
          description: Penjualan berhasil dihapus; response tidak memiliki body
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/penjualan/{id}/status:
    parameters:
      - $ref: '#/components/parameters/PenjualanId'
    patch:
      tags: [Penjualan]
      summary: Ubah status penjualan
      description: |
        Transisi yang diperbolehkan adalah `pending` ke `disetujui` atau
        `dibatalkan`, serta `disetujui` ke `dikirim` atau `dibatalkan`.
      operationId: updateStatusPenjualan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStatusRequest'
      responses:
        '200':
          description: Status berhasil diubah
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PenjualanResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Transisi status tidak diperbolehkan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/ServerError'
components:
  parameters:
    PenjualanId:
      name: id
      in: path
      required: true
      description: ID penjualan
      schema:
        type: integer
        minimum: 1
      example: 1
    Page:
      name: page
      in: query
      description: Nomor halaman
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      description: Jumlah data per halaman
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    StatusFilter:
      name: status
      in: query
      description: Saring daftar berdasarkan status
      schema:
        $ref: '#/components/schemas/StatusPenjualan'
  responses:
    BadRequest:
      description: Request tidak valid atau relasi data tidak ditemukan
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Penjualan tidak ditemukan
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServerError:
      description: Terjadi kesalahan internal pada server
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    StatusPenjualan:
      type: string
      enum: [pending, disetujui, dikirim, dibatalkan]
      example: pending
    Penjualan:
      type: object
      required:
        - id
        - pelanggan_id
        - nama_pelanggan
        - produk_id
        - nama_produk
        - jumlah
        - tanggal
        - total
        - status
      properties:
        id:
          type: integer
          example: 1
        pelanggan_id:
          type: integer
          example: 1
        nama_pelanggan:
          type: string
          nullable: true
          example: Budi Santoso
        produk_id:
          type: integer
          example: 1
        nama_produk:
          type: string
          nullable: true
          example: Kopi Susu
        jumlah:
          type: integer
          minimum: 1
          example: 3
        tanggal:
          type: string
          format: date
          example: '2026-08-01'
        total:
          type: number
          format: double
          minimum: 0
          example: 54000
        status:
          $ref: '#/components/schemas/StatusPenjualan'
    CreatePenjualanRequest:
      type: object
      additionalProperties: false
      required: [pelanggan_id, produk_id, jumlah]
      properties:
        pelanggan_id:
          type: integer
          minimum: 1
          example: 1
        produk_id:
          type: integer
          minimum: 1
          example: 1
        jumlah:
          type: integer
          minimum: 1
          example: 2
        tanggal:
          type: string
          format: date
          description: Jika tidak diberikan, server menggunakan tanggal hari ini.
          example: '2026-09-14'
    UpdateStatusRequest:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/StatusPenjualan'
    Pagination:
      type: object
      required: [page, limit, total, total_pages]
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        total:
          type: integer
          example: 10
        total_pages:
          type: integer
          example: 1
    PenjualanResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/Penjualan'
    PenjualanListResponse:
      type: object
      required: [data, pagination]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Penjualan'
        pagination:
          $ref: '#/components/schemas/Pagination'
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: Penjualan tidak ditemukan